Test FeedItemsMapper in isolation
