+++ EssentialFeed/Feed API/FeedItemsMapper.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-enum FeedItemsMapper {
+public enum FeedItemsMapper {
     private struct Root: Decodable {
         private let items: [RemoteFeedItem]
 
@@ -23,7 +23,7 @@ enum FeedItemsMapper {
         }
     }
 
-    static func map(_ data: Data, from response: HTTPURLResponse) throws -> [FeedImage] {
+    public static func map(_ data: Data, from response: HTTPURLResponse) throws -> [FeedImage] {
         guard response.isOK, let root = try? JSONDecoder().decode(Root.self, from: data)
         else {
             throw RemoteFeedLoader.Error.invalidData
